project_layer/requirements.txt
# Project Layer 依赖（主项目已含的可省略）
pandas>=1.5
numpy>=1.20
scikit-learn>=1.0
statsmodels>=0.13
matplotlib>=3.5
pyyaml>=6.0
joblib>=1.1
pyarrow>=10.0
